+++ src/main/java/model/Calculable.java
@@ -1,11 +1,11 @@
 package model;
 
 public interface Calculable {
-    Integer add(Integer first, Integer second);
-    Integer subtract(Integer first, Integer second);
-    Integer divide(Integer first, Integer second) throws ArithmeticException;
-    Integer multiple(Integer first, Integer second);
+    Double add(Double first, Double second);
+    Double subtract(Double first, Double second);
+    Double divide(Double first, Double second) throws ArithmeticException;
+    Double multiple(Double first, Double second);
 
-    Integer calculate(String input);
-    Integer calculate(Integer first, Integer second, String operator) throws UnknownOperatorException;
+    Double calculate(String input);
+    Double calculate(Double first, Double second, String operator) throws UnknownOperatorException;
 }
