+++ src/main/java/view/Input.java
@@ -1,24 +1,24 @@
 package view;
 
 public class Input {
-    Integer firstNumber;
-    Integer secondNumber;
+    Double firstNumber;
+    Double secondNumber;
     String operator;
 
     public Input() {
     }
 
-    public Input(Integer firstNumber, Integer secondNumber, String operator) {
+    public Input(Double firstNumber, Double secondNumber, String operator) {
         this.firstNumber = firstNumber;
         this.secondNumber = secondNumber;
         this.operator = operator;
     }
 
-    public Integer getFirstNumber() {
+    public Double getFirstNumber() {
         return firstNumber;
     }
 
-    public Integer getSecondNumber() {
+    public Double getSecondNumber() {
         return secondNumber;
     }
